Math Tutor
This program generates addition, subtraction, and multiplication problems for the user to compute.

The program tells the user whether their answer for each problem is correct and prints the results of the entire set as well as each individual problem set. The user can then try another set or stop the program entirely.

For each set of problems, the user customizes the set by telling the program how many problems per set and the maximum number for a number in the problem. Therefore, the user can do 2 addition problems, 5 subtraction problems, and 3 multiplication problems.